pyproject.toml
[tool.poetry]
name = "httprunnerDeom"
version = "0.1.0"
description = ""
authors = ["dubugtalk"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.23.0"
pytest = "^5.3.5"
pyyaml = "^5.3"
jsonpath = "^0.82"

[tool.poetry.dev-dependencies]
coverage = "^5.0.3"
coveralls = "^1.11.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
